requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data processing
librosa>=0.10.0
soundfile>=0.12.0
opencv-python>=4.8.0
Pillow>=10.0.0
h5py>=3.9.0

# Pretrained models (will be downloaded automatically from HuggingFace)
transformers>=4.30.0
fairseq>=0.12.0

# Training utilities
wandb>=0.15.0  # Requires wandb login for experiment tracking
tensorboard>=2.13.0
tqdm>=4.65.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Metrics
lpips>=0.1.4
pytorch-fid>=0.3.0
scikit-image>=0.21.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0
click>=8.1.0
colorlog>=6.7.0
einops>=0.6.0

# Note: Some preprocessing tools may require additional system dependencies
# See data/auspeech_dataset.py for dataset format requirements
